added CommandManager to the list of tools for version 1.0.4
A command manager allow to easily bind the $.__call__() method of a class to different methods depending on the type passed to the argument of the method. For example it becomes very easy to handle dict/list/tuple/str/... differently by using the bind_to(t: type) decorator on the specific method
For now, it doesn't handle inerithance or subclasses for the argument,
also, the __call__ method is forced with a single argument.
To define a default type handling method, just pass the object type to the decorator, the method will then be called when the argument doesn't match any registered type.