Update base.py
quick fix : when data contained Dynamic objects as values, the condition was not evaluated correctly.

Error found when the condition was "not state" and "state" was a DynamicListItem pointing to a False element.
This led to evaluate the condition False instead of True. this is because bool(DynamicListItem(...)) always return True as __bool__ method was not overwritten from object.